requirements.txt
flask
firebase-admin
gunicorn
request 
waitress
